Changes in the following features: Field with render as: :partial:
The Field key in the presenter must not be the partial name anymore, the field key must be the field that will be used to make the data operations (search, order, form, etc.) the partial must be defined in the new parameter called :partial_name . With this change the parameter :related_to was killed.

Search form:

Change the action search to use a submit button intead of a link, to permit use the default submit with enter key in any field of the form.

Change the way Carnival lookups for the label translations to use the active_record translations, so if the models are translated the search form will respect the translations of the model.